+++ expressconnectrouter-20230901/ChangeLog.md
@@ -1,3 +1,8 @@
+2024-03-22 Version: 3.0.0
+- Update API DescribeExpressConnectRouterAllowedPrefixHistory: add param AssociationId.
+- Update API DescribeExpressConnectRouterAllowedPrefixHistory: delete param AssociatonId.
+
+
 2024-03-19 Version: 2.0.0
 - Delete API GetExpressConnectRouter.